chore: moving everthing into dev subdir
