+++ app/dev/blog/[slug]/page.tsx
